move config view from common into redrop
